requirements-dev.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Redis
redis>=5.0.0
aioredis>=2.0.0

# Basic ML (without heavy models for development)
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0

# Security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0